Prevent old data from being processed twice
If the ACK from Receiver never reaches the Sender, Sender will
attempt to resend the same segment identified by the sequence and
segment number.

If the data is a single segment length, the above situation will
cause Receiver to receive the same piece of data twice.

To prevent Receiver from processing the same segment (regarding it
as a valid incoming segment), the receiver will keep track of a list
of recently received segments identified by the source address,
sequence and segment number. If an incoming segment matches any one
of those in the list, it is disregarded.

To prevent the list from building up indefinitely and segment
identifiers not being able to be reused, the list is cleaned at a
regular interval. The entry that is older than a certain amount of
time will be removed.