Add PortRegistry
PortRegistry handles all registerations of port numbers being used
to communicate data via ptp_comms.

It advertises a "register" and "unregister" services to allow callers
to register or unregister port numbers used. Callers must register a
port before using that port to transmit or receive data.

Once a port is registered, a new "rx_data/[port]" topic with [port]
referring to the registered port will be advertised. Callers can then
receive data via that topic.

Callers should also have the etiquette to unregister the port number
if it will not be used anymore in order to free that port number for
future use.